+++ packages/provider_kit/lib/provider_kit.dart
@@ -3,6 +3,7 @@ library provider_kit;
 export 'src/base/index.dart';
 export 'src/cache/index.dart';
 export 'src/core/index.dart';
+export 'src/debounce/index.dart';
 export 'src/errors/index.dart';
 export 'src/mutation/index.dart';
 export 'src/observer/index.dart';